Refactored view
