--- lib/Widgets/slide_item.dart
@@ -1,28 +1,0 @@
-import 'package:flutter/material.dart';
-import 'package:openscan/Utilities/slide.dart';
-
-class SlideItem extends StatelessWidget {
-  final int index;
-
-  SlideItem(this.index);
-
-  @override
-  Widget build(BuildContext context) {
-    Size size = MediaQuery.of(context).size;
-    return Column(
-      mainAxisAlignment: MainAxisAlignment.center,
-      crossAxisAlignment: CrossAxisAlignment.center,
-      children: <Widget>[
-        Container(
-          width: size.width,
-          height: size.height * 0.78,
-          decoration: BoxDecoration(
-            image: DecorationImage(
-              image: AssetImage(slideList[index].imageUrl),
-            ),
-          ),
-        ),
-      ],
-    );
-  }
-}
