+++ lib/presentation/Widgets/gettingStarted/slide_dots.dart
@@ -1,0 +1,22 @@
+import 'package:flutter/material.dart';
+import 'package:openscan/Utilities/constants.dart';
+
+class SlideDots extends StatelessWidget {
+  final bool isActive;
+
+  SlideDots(this.isActive);
+
+  @override
+  Widget build(BuildContext context) {
+    return AnimatedContainer(
+      duration: Duration(milliseconds: 150),
+      margin: const EdgeInsets.symmetric(horizontal: 10),
+      height: isActive ? 12 : 8,
+      width: isActive ? 12 : 8,
+      decoration: BoxDecoration(
+        color: isActive ? secondaryColor : Colors.black,
+        borderRadius: BorderRadius.all(Radius.circular(10)),
+      ),
+    );
+  }
+}
