+++ pubspec.yaml
@@ -11,14 +11,17 @@ environment:
   sdk: ">=2.7.0 <3.0.0"
 
 dependencies:
+  bloc: ^7.0.0
+  cupertino_icons: ^0.1.3
   flutter:
     sdk: flutter
-  cupertino_icons: ^0.1.3
   flutter_absolute_path: ^1.0.6
+  flutter_bloc: ^7.0.1
   flutter_full_pdf_viewer: ^1.0.6
   flutter_speed_dial: ^1.2.5
+  fluttertoast: ^8.0.8
   focused_menu: ^1.0.1
-  image_picker: ^0.6.4
+  image_picker: ^0.8.2
   multi_image_picker: ^4.8.0
   open_file: ^3.0.3
   path_provider: ^2.0.2
